Fix buffer duplication bug
The bug was that a buffer was added to its project-specific
perspective as well as to the perspective it was created in,
i.e. where a call to a project command happend. This fix removes a
buffer from the perspective it was created in, if it does not belong
there.